include: msfw_parport: Fix incorrect type

+++ src/include/msfw_parport.h
@@ -5,6 +5,6 @@
 #define __MSFW_PARPORT_H__
 
 uint16_t msfw_parport_read_byte(void);
-uint8_t msfw_parport_write_byte(uint8_t val);
+uint16_t msfw_parport_write_byte(uint8_t val);
 
 #endif // __MSFW_PARPORT_H__
